+++ source/components/FormField.js
@@ -16,10 +16,10 @@ export type FormFieldProps = {
   disabled?: boolean,
   error?: string | Element<any>,
   isErrorHidden?: boolean,
-  inputRef?: ElementRef<*>,
+  isErrorShown?: boolean,
   label?: string | Element<any>,
   onChange: Function,
-  render: Function,
+  render: (setFormFieldRef: (ElementRef<*>) => void) => React$Node,
   skin?: ComponentType<any>,
   theme: ?Object, // will take precedence over theme in context if passed
   themeId: string,
@@ -41,6 +41,8 @@ class FormFieldBase extends Component<FormFieldProps, State> {
     themeId: IDENTIFIERS.FORM_FIELD,
     themeOverrides: {},
   };
+
+  formFieldRef: ElementRef<*> = React.createRef;
 
   constructor(props: FormFieldProps) {
     super(props);
@@ -66,10 +68,10 @@ class FormFieldBase extends Component<FormFieldProps, State> {
   setError = (error: string) => this.setState({ error });
 
   focusChild = () => {
-    const { inputRef } = this.props;
-    if (inputRef && inputRef.current) {
-      if (typeof inputRef.current.focus === 'function') {
-        inputRef.current.focus();
+    const { formFieldRef } = this;
+    if (formFieldRef && formFieldRef.current) {
+      if (typeof formFieldRef.current.focus === 'function') {
+        formFieldRef.current.focus();
       }
     }
   };
@@ -85,6 +87,7 @@ class FormFieldBase extends Component<FormFieldProps, State> {
         error={error || this.state.error}
         setError={this.setError}
         theme={this.state.composedTheme}
+        formFieldRef={this.formFieldRef}
         focusChild={this.focusChild}
         {...rest}
       />