+++ source/skins/simple/InputSkin.js
@@ -14,15 +14,14 @@ import type { InputProps } from '../../components/Input';
 import { pickDOMProps } from '../../utils/props';
 
 type Props = InputProps & {
-  inputRef: ElementRef<'input'>,
   theme: Object,
   themeId: string,
 };
 
 export const InputSkin = (props: Props) => {
-  const renderInput = () => (
+  const renderInput = (setFormFieldRef) => (
     <input
-      ref={props.inputRef}
+      ref={setFormFieldRef}
       {...pickDOMProps(props)}
       className={classnames([
         props.theme[props.themeId].input,
@@ -35,23 +34,23 @@ export const InputSkin = (props: Props) => {
     />
   );
 
-  const useSelectionRenderer = (option) => (
+  const useSelectionRenderer = (setFormFieldRef, option) => (
     <div className={props.theme[props.themeId].customValueWrapper}>
-      {renderInput()}
+      {renderInput(setFormFieldRef)}
       <div className={props.theme[props.themeId].customValueBlock}>
         {option && props.selectionRenderer && props.selectionRenderer(option)}
       </div>
     </div>
   );
 
-  const render = () => {
+  const render = (setFormFieldRef) => {
     // check if user has passed render prop "selectionRenderer"
     const hasSelectionRenderer =
       props.selectionRenderer && isFunction(props.selectionRenderer);
     if (hasSelectionRenderer) {
-      return useSelectionRenderer(props.selectedOption);
+      return useSelectionRenderer(setFormFieldRef, props.selectedOption);
     }
-    return renderInput();
+    return renderInput(setFormFieldRef);
   };
 
   return (